backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Document parsing
pypdf==3.17.4
pdfplumber==0.11.0
python-docx==1.1.0

# DOCX generation (package name on PyPI is docxtpl)
docxtpl==0.16.7

# AI
google-generativeai>=0.8.0

# Database
supabase==2.3.4

# Validation and settings
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
httpx==0.25.2
